[state-refactoring] Renamed to useApi
